requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
pandas==2.1.3
numpy==1.24.4
plotly==5.17.0
jinja2==3.1.2
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0 